Incorporated GeoTools library to implement ellipsoidal distance between points. Used BigDecimal.

+++ src/main/java/com/dwp/userlocator/city/ProximityCalc.java
@@ -1,6 +1,8 @@
 package com.dwp.userlocator.city;
 
+import java.math.BigDecimal;
+
 public interface ProximityCalc {
     
-    double getMilesBetweenCoordinates(double latitude1, double longitude1, double latitude2, double longitude2);
+    BigDecimal getMilesBetweenCoordinates(double latitude1, double longitude1, double latitude2, double longitude2);
 }
